fastapi/requirements.txt
fastapi
uvicorn
requests
python-dotenv
cohere
